Scheduling-as-a-Service/requirements.txt
Flask==0.12.2
google-cloud-datastore==1.4.0
google-cloud-logging==1.4.0
google-cloud-error_reporting==0.29.0
gunicorn==19.7.1
six==1.11.0
honcho==1.0.1
psq==0.6.0
click==4.1
kubernetes==6.0.0
